Hazards are now in unique rooms
The original code by Dan Malec (and the version by Adafruit which fixed
a compilation issue) had a problem: the screen would show the dangers of
the surrounding rooms by combining them with a bitwise OR, but if
multiple dangers were in the same room, check_for_hazards would only
return one danger.

I could have fixed it by rewriting check_for_hazards but the rest of the
program apparently assumed that there was no more than one hazard in
each room (see for example the switch statement in begin_move_room which
only works correctly if there is never more than one hazard in a room).

So I decided to change the program so that rooms can never have more
than one hazard, which makes it slightly easier to play, and makes the
binary about 140 bytes shorter ;-)